+++ packages/core-cairo/src/api.ts
@@ -1,6 +1,7 @@
 import type { CommonOptions } from './common-options';
 import { printERC20, defaults as erc20defaults, isAccessControlRequired as erc20IsAccessControlRequired, ERC20Options, getInitialSupply } from './erc20';
 import { printERC721, defaults as erc721defaults, isAccessControlRequired as erc721IsAccessControlRequired, ERC721Options } from './erc721';
+import { printERC1155, defaults as erc1155defaults, isAccessControlRequired as erc1155IsAccessControlRequired, ERC1155Options } from './erc1155';
 import { printCustom, defaults as customDefaults, isAccessControlRequired as customIsAccessControlRequired, CustomOptions } from './custom';
 import { toUint256 } from './utils/uint256';
 
@@ -29,6 +30,7 @@ export type ERC20 = WizardContractAPI<ERC20Options> & {
    getInitialSupply: (premint: string, decimals: number) => string;
 }
 export type ERC721 = WizardContractAPI<ERC721Options>;
+export type ERC1155 = WizardContractAPI<ERC1155Options>;
 export type Custom = WizardContractAPI<CustomOptions>;
 
 export const erc20: ERC20 = {
@@ -41,6 +43,11 @@ export const erc721: ERC721 = {
   print: printERC721,
   defaults: erc721defaults,
   isAccessControlRequired: erc721IsAccessControlRequired
+}
+export const erc1155: ERC1155 = {
+  print: printERC1155,
+  defaults: erc1155defaults,
+  isAccessControlRequired: erc1155IsAccessControlRequired
 }
 export const custom: Custom = {
   print: printCustom,